fix: extract rehype codeblock plugin due to lint error
