+++ gatsby-theme-oi-wiki/package.json
@@ -56,6 +56,7 @@
     "rehype-mathjax": "^3.1.0",
     "rehype-parse": "^7.0.1",
     "rehype-pseudocodejs": "file:plugins/rehype-pseudocodejs",
+    "rehype-codeblock": "file:plugins/rehype-codeblock",
     "rehype-react": "^6.2.1",
     "remark-details": "^3.0.0",
     "remark-math": "^3.0.1",
